courseplatform/requirements.txt
asgiref==3.10.0
black==25.9.0
click==8.3.0
cx_Oracle==8.3.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
mssql-django==1.6
mypy_extensions==1.1.0
mysqlclient==2.2.7
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
psycopg2-binary==2.9.11
pyodbc==5.3.0
python-decouple==3.8
pytokens==0.2.0
pytz==2025.2
sqlparse==0.5.3
